Added some more tests for cache/genomeGraph/Parser.
Only one big method "findGenome(source, dest)" in genomeGraph remains to be tested but it is very complex and correlates a lot with the parser. Also this method itself was not very efficient so I think it should be rewrote and made faster using the TOPO sort algorithm Jasper explained last meeting.