Made a new fresh start with Dynamic loading.
Planning on using the old 'enthausiastic' node find algorith since this is not as slow as I thought it was. This does make sure that you have found all the nodes of a layer. (I think?)  Which makes it possible to just add one layer at the time. This can be layout just like the normal layers and added to the drawing.